# Updating stopwatch operations to use singleton