Fixes #
https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/4442

**Changes proposed in this request**
Updating stopwatch operations to use singleton

**Testing**
<!-- Have unit, integration, etc. tests been added? Describe any relevant testing that has been done. -->
<!-- Mention if any and what extra manual testing is needed during the release. -->

**Performance impact**
Negligible impact

Previous benchmark
| Method                | CacheSize   | EnableCacheSerialization | Mean       | Min        | Max        | Gen0    | Gen1    | Allocated |
|---------------------- |------------ |------------------------- |-----------:|-----------:|-----------:|--------:|--------:|----------:|
| AcquireTokenForClient | (1, 10)     | False                    |  10.998 us |  10.525 us |  12.647 us |  0.6714 |       - |  11.36 KB |
| AcquireTokenForClient | (1, 10)     | True                     | 110.139 us | 105.564 us | 118.477 us | 11.4746 |  3.4180 | 188.86 KB |
| AcquireTokenForClient | (10000, 10) | False                    |  18.480 us |  18.381 us |  18.577 us |  0.6714 |       - |  11.37 KB |
| AcquireTokenForClient | (10000, 10) | True                     | 107.464 us | 106.922 us | 108.408 us | 11.4746 | 11.4746 | 188.86 KB |
| AcquireTokenForClient | ?           | ?                        | 223.279 us | 220.400 us | 226.700 us |       - |       - |   37.3 KB |
| AcquireTokenForOBO    | (1, 10)     | False                    |  18.704 us |  18.616 us |  18.803 us |  1.1597 |       - |  19.09 KB |
| AcquireTokenForOBO    | (1, 10)     | True                     | 152.497 us | 151.113 us | 155.036 us | 15.8691 |  5.1270 | 263.12 KB |
| AcquireTokenForOBO    | (10000, 10) | False                    |  28.505 us |  27.649 us |  31.739 us |  1.1597 |       - |   19.1 KB |
| AcquireTokenForOBO    | (10000, 10) | True                     | 152.674 us | 151.935 us | 154.820 us | 15.8691 |  5.1270 | 263.12 KB |
| AcquireTokenForOBO    | ?           | ?                        | 308.000 us | 294.800 us | 319.700 us |       - |       - |  78.63 KB |
| AcquireTokenSilent    | (1, 10)     | ?                        |  15.385 us |  15.272 us |  15.530 us |  0.9766 |       - |  16.43 KB |
| AcquireTokenSilent    | (10000, 10) | ?                        |  25.311 us |  25.222 us |  25.447 us |  0.9766 |       - |  16.44 KB |
| GetAccount            | (1, 10)     | ?                        |   7.921 us |   7.868 us |   7.981 us |  0.4883 |       - |   8.11 KB |
| GetAccount            | (10000, 10) | ?                        |  15.875 us |  15.838 us |  15.919 us |  0.4883 |       - |   8.12 KB |
| RemoveAccount         | (1, 10)     | ?                        |  38.555 us |  37.100 us |  42.000 us |       - |       - |   6.34 KB |
| RemoveAccount         | (10000, 10) | ?                        | 115.380 us | 105.600 us | 142.000 us |       - |       - |   6.36 KB |

New Benchmark

| Method                | CacheSize   | EnableCacheSerialization | Mean       | Min        | Max        | Gen0    | Gen1   | Allocated |
|---------------------- |------------ |------------------------- |-----------:|-----------:|-----------:|--------:|-------:|----------:|
| AcquireTokenForClient | (1, 10)     | False                    |  11.167 us |  11.110 us |  11.311 us |  0.7019 |      - |  11.61 KB |
| AcquireTokenForClient | (1, 10)     | True                     | 108.521 us | 108.134 us | 108.745 us | 11.4746 | 3.4180 | 189.33 KB |
| AcquireTokenForClient | (10000, 10) | False                    |  20.325 us |  20.236 us |  20.386 us |  0.7019 |      - |  11.62 KB |
| AcquireTokenForClient | (10000, 10) | True                     | 108.698 us | 108.216 us | 109.404 us | 11.4746 | 3.0518 | 189.33 KB |
| AcquireTokenForClient | ?           | ?                        | 221.785 us | 214.950 us | 228.050 us |       - |      - |  38.13 KB |
| AcquireTokenForOBO    | (1, 10)     | False                    |  19.495 us |  19.313 us |  19.596 us |  1.1597 |      - |  19.34 KB |
| AcquireTokenForOBO    | (1, 10)     | True                     | 156.097 us | 155.376 us | 157.109 us | 16.1133 | 6.1035 | 263.59 KB |
| AcquireTokenForOBO    | (10000, 10) | False                    |  28.871 us |  28.678 us |  29.161 us |  1.1597 |      - |  19.35 KB |
| AcquireTokenForOBO    | (10000, 10) | True                     | 155.376 us | 154.462 us | 156.940 us | 16.1133 | 4.8828 | 263.59 KB |
| AcquireTokenForOBO    | ?           | ?                        | 310.859 us | 298.900 us | 326.800 us |       - |      - |   79.4 KB |
| AcquireTokenSilent    | (1, 10)     | ?                        |  15.863 us |  15.713 us |  16.027 us |  0.9766 |      - |  16.68 KB |
| AcquireTokenSilent    | (10000, 10) | ?                        |  26.016 us |  25.905 us |  26.181 us |  1.0071 |      - |  16.69 KB |
| GetAccount            | (1, 10)     | ?                        |   7.687 us |   7.647 us |   7.738 us |  0.4883 |      - |   8.11 KB |
| GetAccount            | (10000, 10) | ?                        |  17.546 us |  17.498 us |  17.602 us |  0.4883 |      - |   8.12 KB |
| RemoveAccount         | (1, 10)     | ?                        |  35.580 us |  33.400 us |  37.500 us |       - |      - |   6.34 KB |
| RemoveAccount         | (10000, 10) | ?                        | 123.940 us | 112.000 us | 152.100 us |       - |      - |   6.36 KB |

**Documentation**
- [ ] All relevant documentation is updated.