@bgavrilMS 2024-02-15
Approved as this is from other libraries. But note that the perf testing you've not is not relevant. It's unlikely that such a small change will result in memory / latency increases in an `AcquireTokenForClient` call. A good benchmark would have been to test this in isolation, e.g. 

```csharp
var delay = await MeasureDelay(() => Task.Delay(10));
```

versus the old way of doing things

```cshapr
StopWatch sw = new StopWatch();
sw.Start()
await Task.Delay(10);
sw.ElapsedMilliseconds.
```